+++ src/Products/PloneMeeting/tests/testContacts.py
@@ -475,6 +475,116 @@ class testContacts(PloneMeetingTestCase):
         self.assertEqual(item1.getItemAbsents(), (hp_uid, ))
         self.assertFalse(item2.getItemAbsents())
 
+    def _setup_print_signatories_by_position(self):
+        # Add a position_type
+        self.portal.contacts.position_types += (
+            {"token": u"dg",
+             "name": u"Directeur Général|Directeurs Généraux|"
+                     u"Directrice Générale|Directrices Générales"},
+            {"token": u"super",
+             "name": u"Super-héro|Super-héros|"
+                     u"Super-héroine|Super-héroines"},
+        )
+        person1 = self.portal.contacts.get("person1")
+        person1.firstname = "Jane"
+        person1.lastname = "Doe"
+        person1.gender = u"F"
+        person1.person_title = u"Miss"
+
+        signatory1 = person1.get_held_positions()[0]
+        signatory1.position_type = u"dg"
+        signatory1.secondary_position_type = u"super"
+        signatory1.label = u""
+
+        # No gender/person_title and no position_type, no secondary_position_type
+        person4 = self.portal.contacts.get("person4")
+        person4.firstname = "John"
+        person4.lastname = "Doe"
+        person4.gender = u""
+        person4.person_title = u""
+
+        signatory2 = person4.get_held_positions()[0]
+        signatory2.label = u"Président"
+
+        # prepare 2 more signatories (but not setted as such, yet)
+        person2 = self.portal.contacts.get("person2")
+        signatory3 = person2.get_held_positions()[0]
+        signatory3.label = u"Signatory3"
+
+        person3 = self.portal.contacts.get("person3")
+        signatory4 = person3.get_held_positions()[0]
+        signatory4.label = u"Signatory4"
+
+    def test_pm_print_signatories_by_position(self):
+        self._setup_print_signatories_by_position()
+
+        self.changeUser("pmManager")
+        meeting = self.create("Meeting", date=DateTime())
+        item = self.create("MeetingItem")
+
+        # On MeetingItem
+        view = item.restrictedTraverse("document-generation")
+        helper = view.get_generation_context_helper()
+        # print_signatories_by_position shouldn"t fail if the item is not in a meeting
+        self.assertEqual(len(helper.print_signatories_by_position()), 0)
+        self.presentItem(item)
+
+        printed_signatories = helper.print_signatories_by_position(ender=".")
+        self.assertEqual(
+            printed_signatories,
+            {
+                0: u"La Directrice Générale,",
+                1: u"Jane Doe.",
+                2: u"Président,",  # No position_type, so no prefix
+                3: u"John Doe.",
+            }
+        )
+        printed_signatories = helper.print_signatories_by_position(
+            signature_format=(u"prefixed_secondary_position_type", u"person_with_title", u"XXX", u"gender"),
+            separator=""
+        )
+        self.assertEqual(
+            printed_signatories,
+            {
+                0: u"La Super-héroine",
+                1: u"Miss Jane Doe",
+                2: u"XXX",
+                3: u"F",
+                4: u"Président",  # John Doe has no gender, no title and no secondary_position_type
+                5: u"John Doe",
+                6: u"XXX",
+                7: u""
+            }
+        )
+
+        # On Meeting, with 4 signatories
+        view = meeting.restrictedTraverse("document-generation")
+        helper = view.get_generation_context_helper()
+
+        # Add two more signatories
+        contacts = meeting.orderedContacts.items()
+        signatory3 = contacts[1][1]
+        # Set an absurd value to see if it will be correctly sorted
+        signatory3["signature_number"] = "10"
+        signatory3["signer"] = True
+        signatory4 = contacts[2][1]
+        signatory4["signer"] = True
+        signatory4["signature_number"] = "22"  # Same here
+
+        printed_signatories = helper.print_signatories_by_position(
+            signature_format=(u"position_type",),
+            ender=None
+        )
+        self.assertEqual(
+            printed_signatories,
+            {
+                0: u"Directrice Générale",
+                1: u"Président",
+                2: u"Signatory3",
+                3: u"Signatory4"
+            }
+        )
+
     def _setupInAndOutAttendees(self):
         """Setup a meeting with items and in and out (non) attendees."""
         cfg = self.meetingConfig