+++ CHANGES.rst
@@ -48,11 +48,14 @@ Changelog
   powerobserver for `MeetingConfig` A and not for `MeetingConfig` B and in this case, the available items were shown
 - Added `CKEditor` style `page-break` to be able to insert a `page-break` into a `RichText` field, this can be used in a
   `POD template` by adding a relevant `page-break` paragraph style
+- In `MeetingItemWorkflowConditions._check_review_and_required`, factorized check about `Review portal content` permission and
+  required data (`category/classifier/groupsInCharge`)
+- Improved `BaseDGHV.print_signatories_by_position` to add more use cases
+- Added tests for `BaseDGHV.print_signatories_by_position`
 - Adapted code regarding changes in `collective.iconifiedcategory`, do not use `portal_catalog` to get the annexes but rely on
   `allowedRolesAndUsers` stored in `categorized_elements`
 - Fixed `MeetingView._displayAvailableItemsTo`, do not use `ToolPloneMeeting.userIsAmong` for powerobservers as it could be
   powerobserver for `MeetingConfig` A and not for `MeetingConfig` B and in this case, the available items were shown
-
 
 4.1.27.2 (2020-06-25)
 ---------------------